bug fixes
also manually compiled libxml2 2.9.2.

+++ rework/source/re/math/QuadTree.cpp
@@ -12,7 +12,7 @@
 
 namespace re
 {
-	QuadTree::QuadTree(size_t level, Rect<float, int>& bounds, size_t maxLevels, size_t maxObjects)
+	QuadTree::QuadTree(int level, Rect<float, int>& bounds, int maxLevels, int maxObjects)
 	:m_level(level), m_bounds(bounds), m_maxLevels(maxLevels), m_maxObjects(maxObjects)
 	{
 		for (auto& elem : m_nodes)
@@ -21,7 +21,7 @@ namespace re
 		}
 	}
 
-	QuadTree::QuadTree(size_t level, const Rect<float, int>& bounds, size_t maxLevels, size_t maxObjects)
+	QuadTree::QuadTree(int level, const Rect<float, int>& bounds, int maxLevels, int maxObjects)
 		:m_level(level), m_bounds(bounds), m_maxLevels(maxLevels), m_maxObjects(maxObjects)
 	{
 		for (auto& elem : m_nodes)
